+++ src/main/java/org/opensearch/knn/memoryoptsearch/VectorSearcherFactory.java
@@ -5,6 +5,7 @@
 
 package org.opensearch.knn.memoryoptsearch;
 
+import org.apache.lucene.codecs.hnsw.FlatVectorsScorer;
 import org.apache.lucene.index.FieldInfo;
 import org.apache.lucene.store.Directory;
 import org.apache.lucene.store.IOContext;
@@ -26,5 +27,11 @@ public interface VectorSearcherFactory {
      * @return Null instance if it is not supported, otherwise return {@link VectorSearcher}
      * @throws IOException
      */
-    VectorSearcher createVectorSearcher(Directory directory, String fileName, FieldInfo fieldInfo, IOContext ioContext) throws IOException;
+    VectorSearcher createVectorSearcher(
+        Directory directory,
+        String fileName,
+        FieldInfo fieldInfo,
+        IOContext ioContext,
+        FlatVectorsScorer vectorScorer
+    ) throws IOException;
 }